backend/requirements.txt
# FastAPI backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# 3D geometry processing
trimesh==4.0.5
numpy==1.26.2
scipy==1.11.4

# FDM simulation
pyamg==5.0.1
numba==0.58.1

# Machine Learning
stable-baselines3==2.2.1
gymnasium==0.29.1
torch==2.1.1

# AI/LLM integration
openai==1.3.6
anthropic==0.7.8

# Utilities
pydantic==2.5.2
python-dotenv==1.0.0
aiofiles==23.2.1
